pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mhn-ai-agent-memory"
version = "0.3.0"
description = "AI agent memory system powered by Modern Hopfield Networks — no LLM calls, no database, one matrix multiply"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "shahzebqazi" },
]
keywords = ["hopfield", "modern-hopfield-network", "associative-memory", "attention", "ai-agents", "agent-memory", "context-engineering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.24",
]

[project.optional-dependencies]
semantic = ["sentence-transformers>=2.2"]
openai = ["openai>=1.0"]
scale = ["faiss-cpu>=1.7"]
tfidf = ["scikit-learn>=1.2"]
all = [
    "sentence-transformers>=2.2",
    "openai>=1.0",
    "faiss-cpu>=1.7",
    "scikit-learn>=1.2",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
]

[project.urls]
Repository = "https://github.com/shahzebqazi/mhn-ai-agent-memory"

[tool.hatch.build.targets.wheel]
packages = ["src/hopfield_memory"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]
ignore = ["E501"]
